+++ src/main/kotlin/com/github/gib/CefUrlChangeHandler.kt
@@ -8,6 +8,11 @@ import java.util.function.Consumer
 class CefUrlChangeHandler(private val onUrlChange: Consumer<String?>) : CefDisplayHandlerAdapter() {
 
     override fun onAddressChange(browser: CefBrowser?, frame: CefFrame?, url: String?) {
+        if (url != null && url.startsWith("devtools")) {
+            return
+        }
         onUrlChange.accept(url)
     }
+
+
 }