+++ src/main/kotlin/com/github/gib/SettingsChangedAction.kt
@@ -1,0 +1,12 @@
+package com.github.gib
+
+import com.intellij.util.messages.Topic
+
+interface SettingsChangedAction {
+
+    fun settingsChanged()
+
+    companion object {
+        val TOPIC = Topic.create("Setting Change Topic", SettingsChangedAction::class.java)
+    }
+}